change type n from int to long #1

+++ ft_puthex_lower.c
@@ -6,13 +6,13 @@
 /*   By: lagea < lagea@student.s19.be >             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/25 23:29:19 by lagea             #+#    #+#             */
-/*   Updated: 2024/03/25 23:43:49 by lagea            ###   ########.fr       */
+/*   Updated: 2024/03/26 00:10:18 by lagea            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_puthex_lower(int n)
+void	ft_puthex_lower(long n)
 {
 	char	*base_16;
 	int		index;
