+++ ft_puthex_upper.c
@@ -6,13 +6,13 @@
 /*   By: lagea < lagea@student.s19.be >             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/25 23:45:47 by lagea             #+#    #+#             */
-/*   Updated: 2024/03/25 23:56:50 by lagea            ###   ########.fr       */
+/*   Updated: 2024/03/26 00:10:41 by lagea            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_puthex_upper(int n)
+void	ft_puthex_upper(long n)
 {
 	char	*base_16;
 	int		index;
